# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Dashboard API",
    "description": "RESTful API for managing car rental operations including cars, bookings, customers, expenses, maintenance, and incidents.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token or API key authentication"
      }
    },
    "schemas": {
...
      "Car": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "plate_number": { "type": "string" },
          "make": { "type": "string" },
          "model": { "type": "string" },
          "year": { "type": "integer" },
          "class": { "type": "string" },
          "branch_id": { "type": "integer", "nullable": true },
          "branch_name": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["active", "maintenance", "inactive"] },
          "base_daily_rate": { "type": "number", "format": "float" },
          "vin": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "current_mileage": { "type": "integer" },
          "registration_expiry": { "type": "string", "format": "date", "nullable": true },
          "insurance_expiry": { "type": "string", "format": "date", "nullable": true },
          "doc_count": { "type": "integer" },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "url_path": { "type": "string" }
              }
            }
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "VehicleDocument": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "car_id": { "type": "integer" },
          "document_type": { "type": "string", "enum": ["registration", "insurance", "inspection", "other"] },
          "title": { "type": "string" },
          "expiry_date": { "type": "string", "format": "date", "nullable": true },
          "url_path": { "type": "string" },
          "file_size": { "type": "integer" },
          "uploaded_by": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/cars": {
      "get": {
        "summary": "List all cars",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "maintenance", "inactive"] } },
          { "name": "class", "in": "query", "schema": { "type": "string" } },
          { "name": "branch_id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of cars",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Car" } }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new car",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["plate_number", "make", "model", "year", "class", "base_daily_rate"],
                "properties": {
                  "plate_number": { "type": "string" },
                  "make": { "type": "string" },
                  "model": { "type": "string" },
                  "year": { "type": "integer", "minimum": 1900, "maximum": 2100 },
                  "class": { "type": "string" },
                  "branch_id": { "type": "integer", "nullable": true },
                  "status": { "type": "string", "enum": ["active", "maintenance", "inactive"], "default": "active" },
                  "base_daily_rate": { "type": "number", "minimum": 0 },
                  "vin": { "type": "string", "nullable": true },
                  "notes": { "type": "string", "nullable": true },
                  "current_mileage": { "type": "integer", "minimum": 0 },
                  "registration_expiry": { "type": "string", "format": "date", "nullable": true },
                  "insurance_expiry": { "type": "string", "format": "date", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Car created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Car" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/cars/{id}": {
      "get": {
        "summary": "Get car by ID",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Car details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Car" }
              }
            }
          },
          "404": { "description": "Car not found" }
        }
      },
      "put": {
        "summary": "Update car",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plate_number": { "type": "string" },
                  "make": { "type": "string" },
                  "model": { "type": "string" },
                  "year": { "type": "integer" },
                  "class": { "type": "string" },
                  "branch_id": { "type": "integer", "nullable": true },
                  "status": { "type": "string", "enum": ["active", "maintenance", "inactive"] },
                  "base_daily_rate": { "type": "number", "minimum": 0 },
                  "vin": { "type": "string", "nullable": true },
                  "notes": { "type": "string", "nullable": true },
                  "current_mileage": { "type": "integer", "minimum": 0 },
                  "registration_expiry": { "type": "string", "format": "date", "nullable": true },
                  "insurance_expiry": { "type": "string", "format": "date", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Car updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Car" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Car not found" }
        }
      },
      "delete": {
        "summary": "Archive car (soft delete)",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Car archived",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "403": { "description": "Forbidden" },
          "404": { "description": "Car not found" }
        }
      }
    },
    "/api/cars/{id}/images": {
      "post": {
        "summary": "Upload car images",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" },
                    "maxItems": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "url_path": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/cars/{id}/documents": {
      "get": {
        "summary": "Get car documents",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VehicleDocument" }
                }
              }
            }
          },
          "403": { "description": "Forbidden" }
        }
      },
      "post": {
        "summary": "Upload car document",
        "tags": ["Cars"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["document", "document_type", "title"],
                "properties": {
                  "document": { "type": "string", "format": "binary" },
                  "document_type": { "type": "string", "enum": ["registration", "insurance", "inspection", "other"] },
                  "title": { "type": "string" },
                  "expiry_date": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VehicleDocument" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
...
}